[refactor] #90 OnboardingViewModel 분리
